% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merm.R
\name{graph2mx}
\alias{graph2mx}
\title{Convert a tree to a mapping matrix}
\usage{
graph2mx(td)
}
\arguments{
\item{td}{tree description in mermaid format}
}
\value{
the mapping matrix
}
\description{
Convert a tree description in mermaid format into a mapping matrix
that can be used with the remaining functions in the package.
}
