\name{irtrees-package}
\alias{irtrees-package}
\alias{irtrees}
\docType{package}
\title{Estimation of Tree-Based Item Response Models}
\description{
Helper functions and example data sets to facilitate the estimation 
of tree-based item Response models of the GLMM family with function
\code{glmer} from the \code{lme4} package}
\author{Ivailo Partchev <partchev@gmail.com>}
\references{
  De Boeck, P. and Partchev, I. (2012).
  IRTrees: Tree-Based Item Response Models of the GLMM Family,
  \emph{Journal of Statistical Software -- Code Snippets}, \bold{48}(1), 1--28.
  \url{http://www.jstatsoft.org/v48/c01/}.
}
\keyword{models}
%\examples{}
