% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretree.R
\name{LtoW_single.tree}
\alias{LtoW_single.tree}
\title{Long-to-wide, single tree}
\usage{
LtoW_single.tree(
  data,
  cmx,
  id.col,
  item.col,
  resp.col,
  covar.col = NULL,
  time.col = NULL
)
}
\arguments{
\item{data}{a long-format (person-item-response) data set}

\item{cmx}{a a category-by-node mapping matrix}

\item{id.col}{the person ID column}

\item{item.col}{the item ID column}

\item{resp.col}{the response column}

\item{covar.col}{columns containing covariates}

\item{time.col}{the time column. If not NULL, \code{time.col} 
should be nested in \code{id.col}, since it indicates repeated measures within persons}
}
\description{
Recode a long format data frame to a wide format IRTrees data frame
in the case when items have the same tree structure
}
\details{
Many examples of mapping matrices are given in the vignette. Columns may be specified 
by numeric index or by name. Response categories must be coded with consecutive integers 
starting from 1, not 0. Missing data should be properly represented as NA.
}
