% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretree.R
\name{WtoL_single.tree}
\alias{WtoL_single.tree}
\title{Wide-to-long, one tree}
\usage{
WtoL_single.tree(
  data,
  cmx,
  id.col,
  resp.col,
  covar.col = NULL,
  time.col = NULL
)
}
\arguments{
\item{data}{a wide-format (person-by-item) data set}

\item{cmx}{a a category-by-node mapping matrix}

\item{id.col}{the ID column}

\item{resp.col}{the columns containing the item responses}

\item{covar.col}{columns containing covariates}

\item{time.col}{the time column when there are repeated (longitudinal) data}
}
\description{
Recode a wide format data frame to a long format IRTrees data frame
in the case when all items have the same tree structure
}
\details{
Many examples of mapping matrices are given in the vignette. Columns may be specified 
by numeric index or by name. Response categories must be coded with consecutive integers 
starting from 1, not 0. Missing data should be properly represented as NA.
}
